Bump ssh.test BREE to 1.8.
Enable more warnings and clean them up.

+++ bundles/archived/org.eclipse.equinox.console.ssh.tests/src/org/eclipse/equinox/console/ssh/SshInputScannerTests.java
@@ -56,7 +56,7 @@ public class SshInputScannerTests {
 
         try {
             scanner.scan((byte) 'a');
-            scanner.scan((byte) ESC);
+            scanner.scan(ESC);
             scanner.scan((byte) 'b');
         } catch (IOException e) {
             System.out.println("Error while scanning: " + e.getMessage());
